+++ api/src/main/java/com/polikt/api/news/NewsController.java
@@ -1,16 +1,15 @@
 package com.polikt.api.news;
 
-import org.springframework.http.ResponseEntity;
+import java.util.List;
 
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/news")
@@ -32,7 +31,7 @@ public class NewsController {
 
     // GET /news/{id}
     @GetMapping("/{id}")
-    public ResponseEntity<News> getNewsById(@PathVariable int id) {
+    public ResponseEntity<News> getNewsById(@PathVariable Long id) {
         News news = repository.findById(id).orElse(null);
 
         if (news != null) {
@@ -50,7 +49,7 @@ public class NewsController {
 
     // DELETE /news/{id}
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteNewsById(@PathVariable int id) {
+    public ResponseEntity<Void> deleteNewsById(@PathVariable Long id) {
         if (!repository.existsById(id)) {
             return ResponseEntity.notFound().build();
         }
