+++ api/src/main/java/com/polikt/api/news/NewsRepository.java
@@ -2,5 +2,5 @@ package com.polikt.api.news;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsRepository extends JpaRepository<News, Integer> {
+public interface NewsRepository extends JpaRepository<News, Long> {
 }
